Update
Looks like I can't read rom through a file because the program counter can change to different areas, so next update I'll load it into an array or something.

SDL also seems to have a pain to for it to use window forms so I'll figure out what to do about that later.

Also I will use TextOnly to get a simulated version without using SDL for graphics for test